feat: Default to oldest unsubmitted event on home page
Initialize the selected event on the HomePage to the oldest event for which the user has not yet submitted picks. If picks have been submitted for all events, default to the first event of the season. This improves the user experience by pre-selecting the most relevant upcoming event for pick submission.