feat: Improve UI elements and icons
Updates the styling of the modal in `PicksForm` to use a "carbon-fiber" background and adds a border for better visual definition.

Adjusts the rotation of the F1 car icon in the `LeaderboardPage` to display correctly.

Refactors the `ManageEntitiesPage` to use a new `SaveIcon` component and restyles the save button, integrating it more cleanly with the page layout. Adds the `SaveIcon` component.